widgets: Merge type definition files for mini and regular widgets
