+++ src/components/EditMenu.vue
@@ -311,8 +311,7 @@ import { defaultMiniWidgetManagerVars } from '@/assets/defaults'
 import { MavType } from '@/libs/connection/m2r/messages/mavlink2rest-enum'
 import { isHorizontalScroll } from '@/libs/utils'
 import { useWidgetManagerStore } from '@/stores/widgetManager'
-import { MiniWidgetType } from '@/types/miniWidgets'
-import { type Profile, type View, type Widget, WidgetType } from '@/types/widgets'
+import { type Profile, type View, type Widget, MiniWidgetType, WidgetType } from '@/types/widgets'
 
 import Button from './Button.vue'
 import MiniWidgetInstantiator from './MiniWidgetInstantiator.vue'
